refactor(settings-entry-hooker): merge start-settings intent hooks from SettingsEntryIntentHooker
Move the MainActivity onResume and onNewIntent hooks that consume the start-settings intent flag into SettingsEntryHooker as installLaunchStartSettingsIntentHook and installIncomingStartSettingsIntentHook
