+++ app/src/main/java/io/github/twyora/douyinenhancer/hook/ui/SettingsEntryHooker.kt
@@ -1,6 +1,7 @@
 package io.github.twyora.douyinenhancer.hook.ui
 
 import android.app.Activity
+import android.content.Intent
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
@@ -33,6 +34,8 @@ object SettingsEntryHooker : YukiBaseHooker() {
     override fun onHook() {
         installModuleSettingsEntryHook()
         installAboutAwemeLongClickOpenSettingsHook()
+        installLaunchStartSettingsIntentHook()
+        installIncomingStartSettingsIntentHook()
     }
 
     private fun installModuleSettingsEntryHook(): YukiMemberHookCreator.MemberHookCreator.Result? {
@@ -156,6 +159,70 @@ object SettingsEntryHooker : YukiBaseHooker() {
             }
             onHookingFailure { throwable ->
                 YLog.error("$TAG: failed to hook for attaching long click to open settings", throwable)
+            }
+        }
+    }
+
+    private fun installLaunchStartSettingsIntentHook(): YukiMemberHookCreator.MemberHookCreator.Result? {
+        return packageInstance.mainActivity.selfClass?.resolveMethod(
+            packageInstance.mainActivity.onResume()
+        )?.hook {
+            before {
+                val activity = instance as? Activity ?: return@before
+
+                val shouldStartSettings = activity.intent?.getBooleanExtra(
+                    "douyinenhancer_start_settings",
+                    false
+                )
+                if (shouldStartSettings == true) {
+                    if (verbose) {
+                        YLog.debug("$TAG: start-settings flag detected in onResume intent, showing settings dialog")
+                    }
+                    activity.intent?.removeExtra("douyinenhancer_start_settings")
+                    SettingsDialog.show(activity)
+                    removeSelf {
+                        if (verbose) {
+                            YLog.debug("$TAG: settings dialog shown, unregistering onResume hook to prevent re-show")
+                        }
+                    }
+                }
+            }
+        }?.result {
+            onConductFailure { _, throwable ->
+                YLog.error("$TAG: failed to show settings dialog on resume", throwable)
+            }
+            onHookingFailure { throwable ->
+                YLog.error("$TAG: failed to hook for showing settings dialog on resume", throwable)
+            }
+        }
+    }
+
+    private fun installIncomingStartSettingsIntentHook(): YukiMemberHookCreator.MemberHookCreator.Result? {
+        return packageInstance.mainActivity.selfClass?.resolveMethod(
+            packageInstance.mainActivity.onNewIntent()
+        )?.hook {
+            before {
+                val activity = instance as? Activity ?: return@before
+                val intent = args[0] as? Intent ?: return@before
+
+                val shouldStartSettings = intent.getBooleanExtra(
+                    "douyinenhancer_start_settings",
+                    false
+                )
+                if (shouldStartSettings) {
+                    if (verbose) {
+                        YLog.debug("$TAG: start-settings flag detected in onNewIntent, showing settings dialog")
+                    }
+                    intent.removeExtra("douyinenhancer_start_settings")
+                    SettingsDialog.show(activity)
+                }
+            }
+        }?.result {
+            onConductFailure { _, throwable ->
+                YLog.error("$TAG: failed to show settings dialog on new intent", throwable)
+            }
+            onHookingFailure { throwable ->
+                YLog.error("$TAG: failed to hook for showing settings dialog on new intent", throwable)
             }
         }
     }
